pyproject.toml
[project]
name = "calender-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115.0",
    "google-api-python-client==2.120.0",
    "google-auth==2.28.0",
    "google-auth-httplib2==0.2.0",
    "google-auth-oauthlib==1.2.0",
    "google-genai>=1.2.0",
    "hiredis==2.3.2",
    "httpx==0.27.0",
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-mcp-adapters>=0.1.0",
    "mcp[cli]>=1.0.0",
    "pydantic>=2.7",
    "pydantic-settings>=2.5.2",
    "pytest==8.1.1",
    "pytest-asyncio==0.23.6",
    "python-dateutil==2.9.0",
    "python-dotenv==1.0.1",
    "pytz==2024.1",
    "redis==5.0.3",
    "respx==0.21.1",
    "slowapi==0.1.9",
    "structlog==24.1.0",
    "tenacity==8.2.3",
    "uvicorn[standard]>=0.30.0",
]
